Bayesian Profiling using the Metropolis-Hastings Algorithm

Bayesprofile_MH estimates the probabilities of being included in the target list conditional on unobserved group S using the observed indicator X using the Metropolis-Hastings Algorithm.

Usage

Bayesprofile_MH(dist, mu0, inf0, M, it0, flag_smooth = FALSE)

Arguments

dist

a list containing the distribution information of both the aggregate data and target list (list)

mu0

a vector specifying the mean of the prior distribution (vector)

inf0

a matrix specifying the information matrix of prior distribution (matrix)

M

an integer specifying MCMC iteration numbers (int)

it0

an integer specifying burn-in iteration numbers (int)

flag_smooth

an optional logical variable specifying Dirichlet smoothing or not (bool)

Value

Bayesprofile_MH returns a list containing the following fields:

wmc

the posterior markov chain

nwmc

normalized markov chain

Smc

probabilities of being included in the list in markov chain

loglikmc

log-likelihood of markov chain

rejectmc

the rejection record of the markov chain

dist_plot

The plot of posterior distribution

Details

One thing that users could self-define is to use the Hamiltonian Monte Carlo algorithm (HMC) over the Metropolis-Hastings(MH) method in this function. Generally, the HMC performs better than the MH algorithm: Given the same length of generated chains, the HMC chains usually reaches the steady-state sooner and have a lower log-likelihood. It is hard to incorporate stan in this package, but we gave a block of code in the example to show how to use the HMC algorithm to do Bayesian Profiling, so that the user could run it on their end and do the comparison.

References

De Bruyn, Arnaud, and Thomas Otter (2022) Bayesian Consumer Profiling: How to Estimate Consumer Characteristics from Aggregate Data, Journal of Marketing Research (2022), 59(4), 755–774

Examples

# data preparation
data(VillagesSample)
data(XS_pop)
Election = data_pre(VillagesSample,XS_pop,category = "zipcode_group",base= 'target')
# prior mu = 0, cov = 0.01, M = 4000, burn-in iterations = 200
Result_MH = Bayesprofile_MH(Election, 0, 0.01, 4000, 2000, FALSE)
#> Error in array(x, c(length(x), 1L), if (!is.null(names(x))) list(names(x),     NULL) else NULL): 'data' must be of a vector type, was 'NULL'
# posterior distribution plot
Result_MH$dist_plot
#> Error in eval(expr, envir, enclos): object 'Result_MH' not found

library('rstan')
#> Loading required package: StanHeaders
#> Loading required package: ggplot2
#> rstan (Version 2.21.7, GitRev: 2e1f913d3ca3)
#> For execution on a local, multicore CPU with excess RAM we recommend calling
#> options(mc.cores = parallel::detectCores()).
#> To avoid recompilation of unchanged Stan programs, we recommend calling
#> rstan_options(auto_write = TRUE)
library('reshape2')
library("inline")
library("bayesplot")
#> This is bayesplot version 1.7.2
#> - Online documentation and vignettes at mc-stan.org/bayesplot
#> - bayesplot theme set to bayesplot::theme_default()
#>    * Does _not_ affect other ggplot2 plots
#>    * See ?bayesplot_theme_set for details on theme setting
library("ggplot2")
rstan_options(auto_write = TRUE)
options(mc.cores = parallel::detectCores())

prmu0=rep(0,dists_Election$n_S)
#> Error in eval(expr, envir, enclos): object 'dists_Election' not found
prinf0=diag(dists_Election$n_S)*0.01
#> Error in diag(dists_Election$n_S): object 'dists_Election' not found

data_stan <- list(
  ncatX = dists_Election$n_X,
  ncatS = dists_Election$n_S,
  log_pXcondS = dists_Election$log_pXcondS,
  mu = prmu0,
  sigma = prinf0,
  y = dists_Election$target
)
#> Error in eval(expr, envir, enclos): object 'dists_Election' not found

election.stan = [2172 chars quoted with ''']
# run the stan code
fit <- stan(model_code=election.stan,
            data = data_stan,
            chains = 4,
            iter = 2000)
#> Error in stan(model_code = election.stan, data = data_stan, chains = 4,     iter = 2000): object 'data_stan' not found

posterior <- as.array(fit)
#> Error in as.array(fit): object 'fit' not found

color_scheme_set("red")
my_labels <- c('ABSTENTION','BLANK','SARKOZY','ROYAL','BAYROU','LE PEN','BESANCENOT',
               'VILLIERS','BUFFET','VOYNET','BOVE','LAGUILLER','NIHOUS','SCHIVARDI')
mcmc_intervals(posterior, pars = c("S[1]", "S[2]", "S[14]", "S[10]",
                                   "S[6]", "S[12]", "S[3]", "S[9]",
                                   "S[4]", "S[8]", "S[7]", "S[13]",
                                   "S[11]", "S[5]")) + scale_y_discrete(labels = my_labels)
#> Error in is.data.frame(x): object 'posterior' not found

nw <- extract(fit, 'nw[1]')
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'object' in selecting a method for function 'extract': object 'fit' not found
nw <- unlist(nw, use.names = FALSE)
#> Error in unlist(nw, use.names = FALSE): object 'nw' not found

plot(density(nw),
     xlab=expression(nw), col=grey(0, 0.8),
     main="Parameter distribution")
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'x' in selecting a method for function 'plot': object 'nw' not found